Move file references loader related methods
